dify-batch-processor/dispatcher/requirements.txt
# Google Cloud
google-cloud-datastore
google-cloud-tasks
google-cloud-secret-manager
functions-framework

# For local development .env loading
python-dotenv==0.21.0